Remove duplicated job implementations
