+++ superduper/base/event.py
@@ -1,59 +1,16 @@
-import dataclasses as dc
-import datetime
-import time
 import typing as t
-import uuid
-from abc import ABC, abstractmethod
 
 import numpy
 import pandas
 
 from superduper import logging
-from superduper.backends.base.compute import ComputeBackend
+from superduper.base import Base
 
 if t.TYPE_CHECKING:
     from superduper.base.datalayer import Datalayer
 
 
-@dc.dataclass(kw_only=True)
-class Event(ABC):
-    """Event dataclass to store event data."""
-
-    def dict(self):
-        """Convert to dict."""
-        _base_dict = dc.asdict(self)
-        if 'time' in _base_dict:
-            _base_dict['time'] = str(_base_dict['time'])
-        return {
-            **_base_dict,
-            'genus': self.genus,
-            'queue': self.queue,
-        }
-
-    @classmethod
-    def create(cls, kwargs):
-        """Create event from kwargs.
-
-        :param kwargs: kwargs to create event from
-        """
-        kwargs.pop('genus')
-        kwargs.pop('queue')
-        return cls(**kwargs)
-
-    @abstractmethod
-    def execute(
-        self,
-        db: 'Datalayer',
-    ):
-        """Execute the event.
-
-        :param db: Datalayer instance
-        """
-        pass
-
-
-@dc.dataclass(kw_only=True)
-class Signal(Event):
+class Signal(Base):
     """
     Event used to send a signal to the scheduler.
 
@@ -61,8 +18,8 @@ class Signal(Event):
     :param context: the context of component creation
     """
 
-    genus: t.ClassVar[str] = 'signal'
     queue: t.ClassVar[str] = '_apply'
+
     msg: str
     context: str
 
@@ -78,8 +35,7 @@ class Signal(Event):
             db.cluster.compute.release_futures(self.context)
 
 
-@dc.dataclass(kw_only=True)
-class Change(Event):
+class Change(Base):
     """
     Class for streaming change events.
 
@@ -89,19 +45,9 @@ class Change(Event):
     :param ids: the ids affected
     """
 
-    genus: t.ClassVar[str] = 'change'
     type: str
     queue: str
     ids: t.Sequence[str]
-
-    @classmethod
-    def create(cls, kwargs):
-        """Create event from kwargs.
-
-        :param kwargs: kwargs to create event from
-        """
-        kwargs.pop('genus')
-        return cls(**kwargs)
 
     def execute(
         self,
@@ -114,8 +60,7 @@ class Change(Event):
         raise NotImplementedError('Not relevant for this event class')
 
 
-@dc.dataclass(kw_only=True)
-class Create(Event):
+class Create(Base):
     """
     Class for component creation events.
 
@@ -126,7 +71,6 @@ class Create(Event):
     :param children: the children of the component (if any)
     """
 
-    genus: t.ClassVar[str] = 'create'
     queue: t.ClassVar[str] = '_apply'
 
     context: str
@@ -181,8 +125,7 @@ class Create(Event):
         return f'{self.component}:' f'{self.data["identifier"]}:' f'{self.data["uuid"]}'
 
 
-@dc.dataclass(kw_only=True)
-class Delete(Event):
+class Delete(Base):
     """
     Class for component deletion events.
 
@@ -190,7 +133,6 @@ class Delete(Event):
     :param identifier: the identifier of the component to be deleted
     """
 
-    genus: t.ClassVar[str] = 'delete'
     queue: t.ClassVar[str] = '_apply'
 
     component: str
@@ -236,8 +178,7 @@ class Delete(Event):
         object.cleanup()
 
 
-@dc.dataclass(kw_only=True)
-class Update(Event):
+class Update(Base):
     """
     Update component event.
 
@@ -247,7 +188,6 @@ class Update(Event):
     :param parent: the parent of the component (if any)
     """
 
-    genus: t.ClassVar[str] = 'update'
     queue: t.ClassVar[str] = '_apply'
 
     context: str
@@ -280,128 +220,10 @@ class Update(Event):
         return f'{self.component}' f'{self.data["identifier"]}:' f'{self.data["uuid"]}'
 
 
-@dc.dataclass(kw_only=True)
-class Job(Event):
-    """
-    Job event.
-
-    :param context: context component for job creation
-    :param component: type of component
-    :param identifier: identifier of component
-    :param uuid: uuid of component
-    :param args: arguments of method
-    :param kwargs: kwargs of method
-    :param time: time of job creation
-    :param job_id: id of job
-    :param method: method to run
-    :param status: status of job
-    :param dependencies: list of job_id dependencies
-    """
-
-    genus: t.ClassVar[str] = 'job'
-    queue: t.ClassVar[str] = '_apply'
-
-    context: str
-    component: str
-    identifier: str
-    uuid: str
-    args: t.Sequence[t.Any] = ()
-    kwargs: t.Dict = dc.field(default_factory=dict)
-    time: datetime.datetime = dc.field(default_factory=datetime.datetime.now)
-    job_id: t.Optional[str] = dc.field(default_factory=lambda: str(uuid.uuid4()))
-    method: str
-    status: str = 'pending'
-    dependencies: t.List[str] = dc.field(default_factory=list)
-
-    def get_status(self, db):
-        """Get the status of the job.
-
-        :param db: Datalayer instance
-        """
-        return db['Job'].get(job_id=self.job_id)['status']
-
-    def wait(self, db: 'Datalayer', heartbeat: float = 1, timeout: int = 60):
-        """Wait for the job to finish.
-
-        :param db: Datalayer instance
-        :param heartbeat: time to wait between checks
-        :param timeout: timeout in seconds
-        """
-        start = time.time()
-        status = 'pending'
-        while (status := self.get_status(db)) in {
-            'pending',
-            'running',
-        } and time.time() - start < timeout:
-            if status == 'pending':
-                logging.info(f'Job {self.job_id} is pending')
-            elif status == 'running':
-                logging.info(f'Job {self.job_id} is running')
-            else:
-                break
-
-            time.sleep(heartbeat)
-
-        logging.info(f'Job {self.job_id} finished with status: {status}')
-        return status
-
-    @property
-    def huuid(self):
-        """Return the hashed uuid."""
-        return f'{self.component}:{self.identifier}:{self.uuid}.{self.method}'
-
-    def get_args_kwargs(self, futures):
-        """Get args and kwargs for job execution.
-
-        :param futures: dict of futures
-        """
-        from superduper.backends.base.scheduler import Future
-
-        dependencies = []
-        if self.dependencies:
-            dependencies = [futures[k] for k in self.dependencies if k in futures]
-        args = []
-        for arg in self.args:
-            if isinstance(arg, Future):
-                args.append(futures[arg.job_id])
-            else:
-                args.append(arg)
-        kwargs = {}
-        for k, v in self.kwargs.items():
-            if isinstance(v, Future):
-                kwargs[k] = futures[v.job_id]
-            else:
-                kwargs[k] = v
-        kwargs['dependencies'] = dependencies
-        return args, kwargs
-
-    def execute(
-        self,
-        db: 'Datalayer',
-    ):
-        """Execute the job event.
-
-        :param db: Datalayer instance
-        """
-        meta = {k: v for k, v in self.dict().items() if k not in {'genus', 'queue'}}
-        db.metadata.create_job(meta)
-        return db.cluster.compute.submit(self)
-
-
-events = {
-    'signal': Signal,
-    'change': Change,
-    'update': Update,
-    'create': Create,
-    'job': Job,
-}
-
-
-def unpack_event(dict):
+def unpack_event(r):
     """
     Helper function to deserialize event into Event class.
 
-    :param dict: Serialized event.
+    :param r: Serialized event.
     """
-    event_type = events[dict.get('genus')]
-    return event_type.create(dict)
+    return Base.decode(r)
